Methods getting the error messages from a Symfony form moved from Api ctrl to a trait so it could be used in other ctrls too
