+++ app/bundles/ApiBundle/Controller/CommonApiController.php
@@ -19,6 +19,7 @@ use JMS\Serializer\Exclusion\ExclusionStrategyInterface;
 use JMS\Serializer\SerializationContext;
 use Mautic\ApiBundle\Serializer\Exclusion\ParentChildrenExclusionStrategy;
 use Mautic\ApiBundle\Serializer\Exclusion\PublishDetailsExclusionStrategy;
+use Mautic\CoreBundle\Controller\FormErrorMessagesTrait;
 use Mautic\CoreBundle\Controller\MauticController;
 use Mautic\CoreBundle\Factory\MauticFactory;
 use Mautic\CoreBundle\Form\RequestTrait;
@@ -41,6 +42,7 @@ use Symfony\Component\Translation\TranslatorInterface;
 class CommonApiController extends FOSRestController implements MauticController
 {
     use RequestTrait;
+    use FormErrorMessagesTrait;
 
     /**
      * @var CoreParametersHelper
@@ -460,60 +462,6 @@ class CommonApiController extends FOSRestController implements MauticController
         $this->setSerializationContext($view);
 
         return $this->handleView($view);
-    }
-
-    /**
-     * @param array $formErrors
-     *
-     * @return string
-     */
-    public function getFormErrorMessage(array $formErrors)
-    {
-        $msg = '';
-
-        if ($formErrors) {
-            foreach ($formErrors as $key => $error) {
-                if (!$error) {
-                    continue;
-                }
-
-                if ($msg) {
-                    $msg .= ', ';
-                }
-
-                if (is_string($key)) {
-                    $msg .= $key.': ';
-                }
-
-                if (is_array($error)) {
-                    $msg .= $this->getFormErrorMessage($error);
-                } else {
-                    $msg .= $error;
-                }
-            }
-        }
-
-        return $msg;
-    }
-
-    /**
-     * @param Form $form
-     *
-     * @return array
-     */
-    public function getFormErrorMessages(Form $form)
-    {
-        $errors = [];
-
-        foreach ($form->getErrors(true) as $error) {
-            if (isset($errors[$error->getOrigin()->getName()])) {
-                $errors[$error->getOrigin()->getName()] = [$error->getMessage()];
-            } else {
-                $errors[$error->getOrigin()->getName()][] = $error->getMessage();
-            }
-        }
-
-        return $errors;
     }
 
     /**
